refactor : method of register course & assistant mv to lectureCommandService #24
