+++ swe/src/main/java/com/kwu/swe/domain/lecture/controller/LectureApiController.java
@@ -39,6 +39,28 @@ public class LectureApiController {
         return ApiResponseDto.onSuccess(getLectureResponseDtos(studentLectures));
     }
 
+    @PostMapping("/lectures/{lectureId}")
+    public ApiResponseDto<Long> registerCourse(@PathVariable Long lectureId,
+                                               @RequestParam String studentNumber) {
+        return ApiResponseDto.onSuccess(
+                lectureCommandService.registerCourse(
+                        studentNumber,
+                        lectureId));
+    }
+
+    @PostMapping("/lectures/{lectureId}/assistants/{assistantNumber}")
+    public ApiResponseDto<Long> registerAssistant(@PathVariable Long lectureId,
+                                                  @PathVariable String assistantNumber,
+                                                  @RequestParam String professorNumber) {
+        return ApiResponseDto.onSuccess(
+                lectureCommandService.registerAssistantOfLecture(
+                        professorNumber,
+                        assistantNumber,
+                        lectureId
+                )
+        );
+    }
+
     private static List<LectureResponseDto> getLectureResponseDtos(List<Lecture> allLectures) {
         List<LectureResponseDto> result = allLectures.stream()
                 .map(lecture -> LectureResponseDto.builder()
