+++ swe/src/main/java/com/kwu/swe/domain/lecture/repository/LectureRepository.java
@@ -8,5 +8,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface LectureRepository extends JpaRepository<Lecture, Long> {
-    List<LectureStudent> findByStudent(User student);
 }
