+++ swe/src/main/java/com/kwu/swe/domain/lecture/service/LectureCommandService.java
@@ -6,4 +6,8 @@ public interface LectureCommandService {
 
     //TODO add arguments User entity
     Long registerLecture(String studentNumber, RegisterLectureRequestDto dto);
+
+    Long registerCourse(String studentNumber, Long lectureId);
+
+    Long registerAssistantOfLecture(String professorNumber, String studentNumber, Long lectureId);
 }
