+++ swe/src/main/java/com/kwu/swe/domain/user/repository/LectureStudentRepository.java
@@ -1,7 +1,11 @@
 package com.kwu.swe.domain.user.repository;
 
 import com.kwu.swe.domain.user.entity.LectureStudent;
+import com.kwu.swe.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.List;
+
 public interface LectureStudentRepository extends JpaRepository<LectureStudent, Long> {
+    List<LectureStudent> findByStudent(User student);
 }
